+++ CHANGELOG.md
@@ -4,7 +4,6 @@ Initial Release :
 - Masonry Grid made using List view.
 - ScrollToTop function is added.
 
-
 ## 0.0.2
 
 Bug fix:
@@ -14,7 +13,15 @@ Bug fix:
 
 Bug fix:
 - Tried removing package errors
+
 ## 0.0.4
 
 Bug fix:
 - flutter version downgraded to support pub.dev pana
+
+## 1.0.0
+
+Stable release:
+- Widget name changed from "MasonryGrid" to "MasonryListViewGrid".
+- Documentation update.
+- Example Provided.